Add toggle for area map buffer size
