+++ app/components/map-form.js
@@ -4,6 +4,8 @@ import { service } from '@ember-decorators/service';
 import { argument } from '@ember-decorators/argument';
 import { next } from '@ember/runloop';
 import turfBbox from '@turf/bbox';
+import turfBuffer from '@turf/buffer';
+import turfUnion from '@turf/union';
 import mapboxgl from 'mapbox-gl';
 import projectGeomLayers from '../utils/project-geom-layers';
 
@@ -134,6 +136,8 @@ export default class MapFormComponent extends Component {
 
   paperSize = 'tabloid'
 
+  bufferDistance = 600
+
   paperOrientation = 'landscape'
 
   @argument customLayerGroupQuery = null;
@@ -180,6 +184,28 @@ export default class MapFormComponent extends Component {
       n: `transform: rotate(${360 - bearing}deg)`,
       nSpan: `transform: rotate(${(360 - bearing) * -1}deg)`,
     };
+  }
+
+  // union all geometries together, draw a 600 foot buffer around the union
+  @computed('model.bufferSize')
+  get projectGeometryBuffer() {
+    const geometries = this.getProperties('model.project.developmentSite', 'model.project.projectArea', 'model.project.rezoningArea');
+    const bufferSize = this.get('model.bufferSize');
+    const bufferSizeMiles = bufferSize * 0.000189394;
+
+    const projectGeometryUnion = Object.values(geometries).reduce((union, geometry) => {
+      if (geometry) {
+        if (union === null) {
+          union = geometry;
+        } else {
+          union = turfUnion(union, geometry);
+        }
+      }
+
+      return union;
+    }, null);
+
+    return turfBuffer(projectGeometryUnion, bufferSizeMiles, { units: 'miles' });
   }
 
   @action
@@ -280,7 +306,7 @@ export default class MapFormComponent extends Component {
 
   @action
   fitBoundsToBuffer() {
-    const buffer = this.get('model.project.projectGeometryBuffer');
+    const buffer = this.get('projectGeometryBuffer');
     const map = this.get('mapInstance');
 
     map.setBearing(0);
