+++ app/models/area-map.js
@@ -2,4 +2,6 @@ import ApplicantMap from './applicant-map';
 
 export default class extends ApplicantMap {
   mapTypeLabel = 'Area Map';
+
+  bufferSize = '600';
 }
