+++ app/templates/projects/edit/map/edit.hbs
@@ -10,7 +10,7 @@
     {{#if model.project.rezoningArea}}
       {{labs-ui/legend-item item=areaMapLegendConfig.rezoningArea}}
     {{/if}}
-    {{#if model.project.projectGeometryBuffer}}
+    {{#if model.projectGeometryBuffer}}
       {{labs-ui/legend-item item=areaMapLegendConfig.projectGeometryBuffer}}
     {{/if}}
   </section>
