+++ app/templates/projects/edit/map/new.hbs
@@ -18,7 +18,7 @@
           {{labs-ui/legend-item item=areaMapLegendConfig.rezoningArea}}
         </div>
       {{/if}}
-      {{#if model.project.projectGeometryBuffer}}
+      {{#if model.projectGeometryBuffer}}
         <div class="cell project-geom-cell">
           {{labs-ui/legend-item item=areaMapLegendConfig.projectGeometryBuffer}}
         </div>
